requirements.txt
# =============================================================================
# DSH ETL Search AI - Requirements
# =============================================================================
# Lightweight setup using Cohere API for embeddings (no torch/scipy needed)
# Total install size: ~80MB
# =============================================================================

# -----------------------------------------------------------------------------
# Core
# -----------------------------------------------------------------------------
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# -----------------------------------------------------------------------------
# HTTP & Async
# -----------------------------------------------------------------------------
aiohttp>=3.9.0      # Async HTTP client (used in http.py)
aiofiles>=23.0.0    # Async file I/O (used in local.py, cached.py)
httpx>=0.25.0       # Alternative HTTP client, useful for sync contexts

# -----------------------------------------------------------------------------
# XML/JSON Parsing
# -----------------------------------------------------------------------------
lxml>=4.9.0
rdflib>=7.0.0

# -----------------------------------------------------------------------------
# Database
# -----------------------------------------------------------------------------
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# -----------------------------------------------------------------------------
# Vector Operations (lightweight)
# -----------------------------------------------------------------------------
numpy>=1.24.0

# -----------------------------------------------------------------------------
# Embeddings (Cohere API)
# -----------------------------------------------------------------------------
cohere>=5.0.0

# -----------------------------------------------------------------------------
# PDF Processing
# -----------------------------------------------------------------------------
pypdf>=3.0.0

# -----------------------------------------------------------------------------
# API Framework
# -----------------------------------------------------------------------------
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# -----------------------------------------------------------------------------
# Testing
# -----------------------------------------------------------------------------
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# -----------------------------------------------------------------------------
# Logging
# -----------------------------------------------------------------------------
structlog>=24.0.0

# -----------------------------------------------------------------------------
# Development
# -----------------------------------------------------------------------------
black>=23.0.0
isort>=5.12.0email-validator>=2.0.0
chromadb>=0.4.0
cohere>=4.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
email-validator>=2.0.0
